+++ modules/instance_template/variables.tf
@@ -168,14 +168,11 @@ variable "shielded_instance_config" {
 ###########################
 # Public IP
 ###########################
-variable "public_ip" {
-  description = "Determines whether a public IP address is added to your VM instance."
-  type        = bool
-  default     = false
-}
-
 variable "access_config" {
   description = "Access configurations, i.e. IPs via which the VM instance can be accessed via the Internet."
-  type        = map(string)
-  default     = {}
+  type        = list(object({
+    nat_ip       = string
+    network_tier = string
+  }))
+  default     = []
 }