+++ modules/preemptible_and_regular_instance_templates/variables.tf
@@ -139,14 +139,11 @@ variable "service_account" {
 ###########################
 # Public IP
 ###########################
-variable "public_ip" {
-  description = "Determines whether a public IP address is added to your VM instance."
-  type        = bool
-  default     = false
-}
-
 variable "access_config" {
   description = "Access configurations, i.e. IPs via which the VM instance can be accessed via the Internet."
-  type        = map(string)
-  default     = {}
+  type        = list(object({
+    nat_ip       = string
+    network_tier = string
+  }))
+  default     = []
 }